Implemented Custom Cities context to make few states globally available
