+++ 15-Worldwise/src/components/CityList.jsx
@@ -2,8 +2,11 @@ import Spinner from './Spinner'
 import CityItem from './CityItem'
 import Message from './Message'
 import styles from './CityList.module.css'
+import { useCities } from '../contexts/CitiesContext'
 
-function CityList({ cities, isLoading }) {
+function CityList() {
+  const { cities, isLoading } = useCities()
+
   if (isLoading) return <Spinner />
 
   if (!cities.length)
