+++ 15-Worldwise/src/components/CountryList.jsx
@@ -2,8 +2,11 @@ import Spinner from './Spinner'
 import Message from './Message'
 import CountryItem from './CountryItem'
 import styles from './CountryList.module.css'
+import { useCities } from '../contexts/CitiesContext'
 
-function CountryList({ cities, isLoading }) {
+function CountryList() {
+  const { cities, isLoading } = useCities()
+
   if (isLoading) return <Spinner />
 
   if (!cities.length)
