+++ 15-Worldwise/src/contexts/CitiesContext.jsx
@@ -1,0 +1,42 @@
+import { createContext, useContext, useEffect, useState } from 'react'
+
+const BASE_URL = 'http://localhost:9000'
+
+const CitiesContext = createContext()
+
+function CitiesProvider({ children }) {
+  const [cities, setCities] = useState([])
+  const [isLoading, setIsLoading] = useState(false)
+
+  useEffect(function () {
+    async function fetchCities() {
+      try {
+        const res = await fetch(`${BASE_URL}/cities`)
+        const data = await res.json()
+        setCities(data)
+      } catch {
+        alert('There was an error loading data...')
+      } finally {
+        setIsLoading(false)
+      }
+    }
+
+    fetchCities()
+  }, [])
+
+  return (
+    <CitiesContext.Provider value={{ cities, isLoading }}>
+      {children}
+    </CitiesContext.Provider>
+  )
+}
+
+function useCities() {
+  const context = useContext(CitiesContext)
+  if (context === undefined)
+    throw new Error('CitiesContext was used outside the CitiesProvider')
+
+  return context
+}
+
+export { CitiesProvider, useCities }
